Updated procurement, and order details

+++ backend/API/Services/Procurements/ProcurementService.cs
@@ -24,20 +24,40 @@ namespace API.Services.Procurements
 
         public async Task<ProcurementResDto> CreateProcurement(ProcurementReqDto procurementReqDto)
         {
-            var procurementCategory = await _context.ProductCategories
+            var createdBy = await _context.Users
+                .Include(x => x.Role)
+                .FirstOrDefaultAsync(x => x.Id == procurementReqDto.CreatedById);
+
+            if (createdBy == null)
+                throw new NotFoundException("User not found.");
+
+            if (!UserRoleEnum.ProcurementMakerRoles.Contains(createdBy.Role.Name))
+                throw new BadRequestException("User is not allowed to create procurement.");
+
+            var procurementCategory = await _context.Categories
                 .Where(x => x.Id == procurementReqDto.ProcurementCategoryId)
                 .FirstOrDefaultAsync();
 
             if (procurementCategory == null)
                 throw new NotFoundException("Procurement category not found.");
 
+            var estimatedTotalPrice = procurementReqDto.Products.Sum(
+                x => x.EstimatedPrice * x.Quantity
+            );
+
+            if (estimatedTotalPrice > 300000)
+                throw new BadRequestException(
+                    "Estimated subtotal price cannot be greater than 300,000."
+                );
+
             var procurement = new Procurement
             {
+                CreatedBy = createdBy,
                 Title = procurementReqDto.Title,
                 Category = procurementCategory,
-                EstimatedTotalPrice = procurementReqDto.EstimatedTotalPrice,
+                EstimatedTotalPrice = estimatedTotalPrice,
                 Deadline = procurementReqDto.Deadline,
-                Status = StatusEnum.NoOffer,
+                Status = StatusEnum.PendingApproval,
                 Products = procurementReqDto.Products.ConvertAll<ProcurementProduct>(x =>
                 {
                     return new ProcurementProduct
@@ -68,6 +88,11 @@ namespace API.Services.Procurements
 
             if (procurement == null)
                 throw new NotFoundException("Procurement not found.");
+            if (procurement.Status == StatusEnum.OfferAccepted)
+                throw new ApiException(
+                    HttpStatusCode.Forbidden,
+                    "Can't delete procurement after accepting an offer."
+                );
 
             _context.Procurements.Remove(procurement);
 
@@ -87,6 +112,8 @@ namespace API.Services.Procurements
                 .Include(x => x.Quotations)
                 .ThenInclude(x => x.Supplier)
                 .ThenInclude(s => s.Category)
+                .Include(x => x.CreatedBy)
+                .ThenInclude(x => x.Role)
                 .FirstOrDefaultAsync();
 
             if (procurement == null)
@@ -95,28 +122,19 @@ namespace API.Services.Procurements
             return _mapper.Map<ProcurementResDto>(procurement);
         }
 
-        public async Task<List<ProcurementResDto>> GetProcurements(
-            ProcurementsGetParams procurementsGetParams
-        )
+        public async Task<List<ProcurementResDto>> GetProcurements()
         {
-            var procurementList = new List<Procurement>();
-
-            IQueryable<Procurement> procurements = _context.Procurements.Include(x => x.Category);
-            if (procurementsGetParams?.CategoryId > 0)
-            {
-                procurements = procurements.Where(
-                    x => x.Category.Id == procurementsGetParams.CategoryId
-                );
-            }
-
-            procurementList = await procurements
+            var procurements = await _context.Procurements
+                .Include(x => x.Category)
+                .Include(x => x.Products)
+                .ThenInclude(x => x.Product)
+                .ThenInclude(x => x.Category)
                 .Include(x => x.Quotations)
                 .ThenInclude(x => x.Supplier)
+                .ThenInclude(s => s.Category)
                 .ToListAsync();
 
-            var procurementListResDto = _mapper.Map<List<ProcurementResDto>>(procurementList);
-
-            return _mapper.Map<List<ProcurementResDto>>(procurementList);
+            return _mapper.Map<List<ProcurementResDto>>(procurements);
         }
     }
 }
